Title: 괄호 쌍 만들어주기, Time: 8ms, Memory: 1MB, Status: Passed - CodeTree

+++ 240410/괄호 쌍 만들어주기/pair-parentheses.cpp
@@ -2,9 +2,9 @@
 #include <vector>
 using namespace std;
 
-int arr[100005];
+long long arr[100005];
 string s;
-int ans=0;
+long long ans=0;
 
 int main() {
